Delete old implement and update Vue test result
