+++ tests/Signal/subtle/watch-unwatch.test.ts
@@ -1,5 +1,5 @@
 import {describe, expect, it, vi} from 'vitest';
-import {Signal} from '../../../src/alien.js';
+import {Signal} from '../../../src';
 
 describe('watch and unwatch', () => {
   it('handles multiple watchers well', () => {
