+++ tests/behaviors/liveness.test.ts
@@ -1,5 +1,5 @@
 import {describe, expect, it, vi} from 'vitest';
-import {Signal} from '../../src/alien.js';
+import {Signal} from '../../src';
 
 describe('liveness', () => {
   it('only changes on first and last descendant', () => {
