+++ tests/behaviors/prohibited-contexts.test.ts
@@ -1,5 +1,5 @@
 import {describe, expect, it} from 'vitest';
-import {Signal} from '../../src/alien.js';
+import {Signal} from '../../src';
 
 describe('Prohibited contexts', () => {
   it('allows writes during computed', () => {
