+++ tests/behaviors/type-checking.test.ts
@@ -1,5 +1,5 @@
 import {describe, expect, it} from 'vitest';
-import {Signal} from '../../src/alien.js';
+import {Signal} from '../../src';
 
 describe('Expected class shape', () => {
   it('should be on the prototype', () => {
